Move about view to fourth tab
This is a prep commit for adding a picker on the home view
that displays a list of elections to choose from. Thus, we
no longer need the more elections tab. It was replaced with
the about tab to make more space on the home view for the
extra picker.

Also:
Remove unnecessary NavigationController for ContactsSearchVC
Remove FindElections controller source