Add election picker to Home View
Required refactor of Election to split into a
UserElection and Election, where UserElection contains the
additional userAddress property and all the voterInfoQuery data.

UserElection is a sub-entity of Election in the CoreData model.
This has the side-effect of requiring us to filter fetch results.

Election is simply the CoreData representation of the electionInfo
query of the Civic API.