Bugfix #117: Use telprompt url scheme on candidate call action
The telprompt scheme does the following:
  - Displays an AlertView asking the user to confirm the call
  - Returns the user to the calling app after the phone call
    is completed

This telprompt url scheme is undocumented, but is used regularly
by other developers and there has never been a reported case of
Apple rejecting an app for its use. This url scheme is used
internally by UIWebView and MobileSafari.

If this url scheme is rejected or unsupported, we could easily
mimic its confirm functionality with an AlertView. The return user
to calling app functionality would be more difficult.