Revert and lock to GA 3.0.7
Google Analytics as of 3.0.8 defaults to collecting
the IDFA of each device and throws a warning that
the required idfa lib is not linked on app start.

Apple rejects apps that use IDFA without serving ads,
so for now we revert to a version that does not use IDFA
to avoid potential app store issues.

In the future we could specify turning off idfa collection via
the GA SDK, and attempt a submission to the app store. For now,
we aren't risking it since we want this update to go out in a
timely fashion.