Clear address field
Whenever starting the app or going to the home page (ContactsSearchViewController) from within the app, ensure the address field is cleared so a new address is entered every time and an API request is called with that address prior to continuing. Do not save information from the API call in user defaults to be used later, make a fresh API call on every address change.